refactor(cli): split arg grammar out of manifest-cli into manifest-cli-args (#836)
manifest-cli.ts (850 lines) and its test (937 lines) exceeded the repo's
~400-line guideline after #691's first CLI-layer split. Introduce the second
seam #836 calls for: flag parsing, flag-combination validation, and --help
text move into manifest-cli-args.ts, consumed by manifest-cli.ts, which keeps
the check/emit execution. Tests split the same way. Pure move, no behavior
change; gaia-maintainer binary rebuilt to match (gaia adopter binary and
templates are byte-identical since manifest-cli.ts is maintainer-only).
